requirements/requirements.txt
Django==1.6
django-extensions==1.3.2
django-choices==1.1.12
django-author==0.1
distribute==0.6.24
MySQL-python==1.2.5
django-bootstrap3==2.5.5
